+++ ShipEngine/ShipEngine.cs
@@ -363,5 +363,123 @@ namespace ShipEngineSDK
 
             return labelResult;
         }
+
+        /// <summary>
+        /// Update a shipment from shipment details
+        /// </summary>
+        /// <param name="shipmentParams">Details of the shipment that you want to update</param>
+        /// <param name="shipmentID">Shipment ID to update</param>
+        /// <returns>Object containing the updated shipment information</returns>
+        public async Task<UpdateShipmentByID.Result> UpdateShipmentFromShipmentDetails(UpdateShipmentByID.Params shipmentParams, string shipmentID)
+        {
+
+            string shipmentParamsString = JsonConvert.SerializeObject(shipmentParams.Shipment, JsonSerializerSettings);
+
+            var path = $"/v1/shipments/{shipmentID}";
+
+            var labelResult = await SendHttpRequestAsync<UpdateShipmentByID.Result>(HttpMethod.Put, path, shipmentParamsString, _client, _config);
+
+            return labelResult;
+        }
+
+        /// <summary>
+        /// Update a shipment from shipment details
+        /// </summary>
+        /// <param name="shipmentParams">Details of the shipment that you want to update</param>
+        /// <param name="shipmentID">Shipment ID to update</param>
+        /// <param name="methodConfig">Configuration object that overrides the global config for this method call</param>
+        /// <returns>Object containing the updated shipment information</returns>
+        public async Task<UpdateShipmentByID.Result> UpdateShipmentFromShipmentDetails(UpdateShipmentByID.Params shipmentParams, string shipmentID, Config methodConfig)
+        {
+
+            var client = ConfigureHttpClient(methodConfig, new HttpClient());
+
+            string shipmentParamsString = JsonConvert.SerializeObject(shipmentParams.Shipment, JsonSerializerSettings);
+
+            var path = $"/v1/shipments/{shipmentID}";
+
+            var labelResult = await SendHttpRequestAsync<UpdateShipmentByID.Result>(HttpMethod.Put, path, shipmentParamsString, client, methodConfig);
+
+            client.Dispose();
+
+            return labelResult;
+        }
+
+        /// <summary>
+        /// Create one or multiple shipments
+        /// </summary>
+        /// <param name="shipmentParams">List of details of the shipments that you want to create</param>
+        /// <returns>Object containing the created shipments information</returns>
+        public async Task<CreateShipments.Result> CreateShipments(CreateShipments.Params shipmentParams)
+        {
+
+            string shipmentParamsString = JsonConvert.SerializeObject(shipmentParams.Shipments, JsonSerializerSettings);
+
+            var path = $"/v1/shipments";
+
+            var shipmentResult = await SendHttpRequestAsync<CreateShipments.Result>(HttpMethod.Post, path, shipmentParamsString, _client, _config);
+
+            return shipmentResult;
+        }
+
+        /// <summary>
+        /// Create one or multiple shipments
+        /// </summary>
+        /// <param name="shipmentParams">List of details of the shipments that you want to create</param>
+        /// <param name="methodConfig">Configuration object that overrides the global config for this method call</param>
+        /// <returns>Object containing the created shipments information</returns>
+        public async Task<CreateShipments.Result> CreateShipments(CreateShipments.Params shipmentParams, Config methodConfig)
+        {
+
+            var client = ConfigureHttpClient(methodConfig, new HttpClient());
+
+            string shipmentParamsString = JsonConvert.SerializeObject(shipmentParams.Shipments, JsonSerializerSettings);
+
+            var path = $"/v1/shipments";
+
+            var shipmentResult = await SendHttpRequestAsync<CreateShipments.Result>(HttpMethod.Post, path, shipmentParamsString, _client, _config);
+
+            client.Dispose();
+
+            return shipmentResult;
+        }
+
+        /// <summary>
+        /// Create a label from a rate id
+        /// </summary>
+        /// <param name="createLabelFromShipmentIDParams">The details of the shipment that you want to use to purchase a label</param>
+        /// <returns>Object containing the created label information</returns>
+        public async Task<CreateLabelFromShipmentID.Result> CreateLabelFromShipmentID(CreateLabelFromShipmentID.Params createLabelFromShipmentIDParams)
+        {
+            var path = $"/v1/labels/shipment/{createLabelFromShipmentIDParams.ShipmentID}";
+
+            string createLabelFromShipmentIDParamsString = JsonConvert.SerializeObject(createLabelFromShipmentIDParams, JsonSerializerSettings);
+
+            var labelResult = await SendHttpRequestAsync<CreateLabelFromShipmentID.Result>(HttpMethod.Post, path, createLabelFromShipmentIDParamsString, _client, _config);
+
+            return labelResult;
+        }
+
+        /// <summary>
+        /// Create a label from a rate id
+        /// </summary>
+        /// <param name="createLabelFromShipmentIDParams">The details of the shipment that you want to use to purchase a label</param>
+        /// <param name="methodConfig">Configuration object that overrides the global config for this method call</param>
+        /// <returns>Object containing the created label information</returns>
+        public async Task<CreateLabelFromShipmentID.Result> CreateLabelFromShipmentID(CreateLabelFromShipmentID.Params createLabelFromShipmentIDParams, Config methodConfig)
+        {
+
+            var client = ConfigureHttpClient(methodConfig, new HttpClient());
+
+            var path = $"/v1/labels/shipment/{createLabelFromShipmentIDParams.ShipmentID}";
+
+            string createLabelFromShipmentIDParamsString = JsonConvert.SerializeObject(createLabelFromShipmentIDParams, JsonSerializerSettings);
+
+            var labelResult = await SendHttpRequestAsync<CreateLabelFromShipmentID.Result>(HttpMethod.Post, path, createLabelFromShipmentIDParamsString, client, methodConfig);
+
+            client.Dispose();
+
+            return labelResult;
+        }
     }
 }